+++ index.js
@@ -49,6 +49,10 @@ class PSPDFKitView extends React.Component {
   /**
    * @ignore
    */
+  _notificationCenter = null;
+  /**
+   * @ignore
+   */
   _componentRef = React.createRef(this);
 
   render() {
@@ -76,6 +80,7 @@ class PSPDFKitView extends React.Component {
           onCustomToolbarButtonTapped={this._onCustomToolbarButtonTapped}
           onCustomAnnotationContextualMenuItemTapped={this._onCustomAnnotationContextualMenuItemTapped}
           onDocumentScrolled={this._onDocumentScrolled}
+         
         />
       );
     } else {
@@ -191,10 +196,10 @@ class PSPDFKitView extends React.Component {
    * @ignore
    */
   _onCustomAnnotationContextualMenuItemTapped = event => {
-      if (this.props.onCustomAnnotationContextualMenuItemTapped) {
-        this.props.onCustomAnnotationContextualMenuItemTapped(event.nativeEvent);
-      }
-    };
+    if (this.props.onCustomAnnotationContextualMenuItemTapped) {
+      this.props.onCustomAnnotationContextualMenuItemTapped(event.nativeEvent);
+    }
+  };
 
   /**
    * Enters annotation creation mode, showing the annotation creation toolbar.
@@ -271,7 +276,7 @@ class PSPDFKitView extends React.Component {
 
   /**
    * Saves the document that’s currently open.
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.save()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().save()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.save|save()}.
    * @method saveCurrentDocument
    * @memberof PSPDFKitView
@@ -324,8 +329,26 @@ class PSPDFKitView extends React.Component {
   };
 
   /**
+   * Get the current Notification Center.
+   * @method getNotificationCenter
+   * @example
+   * const document = this.pdfRef.current?.getNotificationCenter();
+   * @see {@link https://pspdfkit.com/api/react-native/NotificationCenter.html} for available methods.
+   * @memberof PSPDFKitView
+   * @returns { NotificationCenter } A reference to the Notification Center that can be used to subscribe and unsubscribe from events.
+   */
+  getNotificationCenter () {
+    if (this._notificationCenter == null) {
+      this._notificationCenter = new NotificationCenter(this._componentRef.current);
+      return this._notificationCenter;
+    } else {
+      return this._notificationCenter;
+    }
+  };
+
+  /**
    * @method clearSelectedAnnotations
-   * @memberof PDFDocument
+   * @memberof PSPDFKitView
    * @description Clears all currently selected Annotations.
    * @example
    * const result = await this.pdfRef.current?.clearSelectedAnnotations();
@@ -356,15 +379,15 @@ class PSPDFKitView extends React.Component {
     }
   };
 
- /**
-  * @method selectAnnotations
-  * @memberof PDFDocument
-  * @param { object } annotations An array of the annotations to select in Instant JSON format.
-  * @description Select one or more annotations.
-  * @example
-  * const result = await this.pdfRef.current?.selectAnnotations(annotations);
-  * @returns { Promise<any> } A promise containing the result of the operation. ```true``` if the annotations were selected, ```false``` otherwise.
-  */
+  /**
+   * @method selectAnnotations
+   * @memberof PSPDFKitView
+   * @param { object } annotations An array of the annotations to select in Instant JSON format.
+   * @description Select one or more annotations.
+   * @example
+   * const result = await this.pdfRef.current?.selectAnnotations(annotations);
+   * @returns { Promise<any> } A promise containing the result of the operation. ```true``` if the annotations were selected, ```false``` otherwise.
+   */
   selectAnnotations = function (annotations) {
     if (Platform.OS === 'android') {
       let requestId = this._nextRequestId++;
@@ -395,7 +418,7 @@ class PSPDFKitView extends React.Component {
    * Gets all annotations of the given type from the specified page.
    *
    * @method getAnnotations
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.getAnnotations()``` or ```getAnnotationsForPage()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().getAnnotations()``` or ```getAnnotationsForPage()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.getAnnotations|getAnnotations()} and {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.getAnnotationsForPage|getAnnotationsForPage()}.
    * @memberof PSPDFKitView
    * @param { number } pageIndex The page index to get the annotations for, starting at 0.
@@ -436,7 +459,7 @@ class PSPDFKitView extends React.Component {
    * Adds a new annotation to the current document.
    *
    * @method addAnnotation
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.addAnnotations()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().addAnnotations()``` instead.
    * @memberof PSPDFKitView
    * @param { object } annotation The InstantJSON of the annotation to add.
    * @example
@@ -474,7 +497,7 @@ class PSPDFKitView extends React.Component {
    * Removes an existing annotation from the current document.
    *
    * @method removeAnnotation
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.removeAnnotations()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().removeAnnotations()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.removeAnnotations|removeAnnotations()}.
    * @memberof PSPDFKitView
    * @param { object } annotation The InstantJSON of the annotation to remove.
@@ -512,7 +535,7 @@ class PSPDFKitView extends React.Component {
    * Removes the supplied document InstantJSON from the current document.
    *
    * @method removeAnnotations
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.removeAnnotations()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().removeAnnotations()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.removeAnnotations|removeAnnotations()}.
    * @memberof PSPDFKitView
    * @param { object } annotation The InstantJSON of the annotations to remove.
@@ -550,7 +573,7 @@ class PSPDFKitView extends React.Component {
    * Gets all unsaved changes to annotations.
    *
    * @method getAllUnsavedAnnotations
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.getAllUnsavedAnnotations()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().getAllUnsavedAnnotations()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.getAllUnsavedAnnotations|getAllUnsavedAnnotations()}.
    * @memberof PSPDFKitView
    * @returns { Promise } A promise containing document InstantJSON.
@@ -585,7 +608,7 @@ class PSPDFKitView extends React.Component {
    * Gets all annotations of the given type.
    *
    * @method getAllAnnotations
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.getAnnotations()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().getAnnotations()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.getAnnotations|getAnnotations()}.
    * @memberof PSPDFKitView
    * @param { string } [type] The type of annotations to get. If not specified or ```null```, all annotation types will be returned.
@@ -626,7 +649,7 @@ class PSPDFKitView extends React.Component {
    * Applies the supplied document InstantJSON to the current document.
    *
    * @method addAnnotations
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.addAnnotations()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().addAnnotations()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.addAnnotations|addAnnotations()}.
    * @memberof PSPDFKitView
    * @param { object } annotations The document InstantJSON to apply to the current document.
@@ -685,7 +708,8 @@ class PSPDFKitView extends React.Component {
 
       UIManager.dispatchViewManagerCommand(
         findNodeHandle(this._componentRef.current),
-        this._getViewManagerConfig('RCTPSPDFKitView').Commands.setAnnotationFlags,
+        this._getViewManagerConfig('RCTPSPDFKitView').Commands
+          .setAnnotationFlags,
         [requestId, uuid, flags],
       );
 
@@ -722,7 +746,8 @@ class PSPDFKitView extends React.Component {
 
       UIManager.dispatchViewManagerCommand(
         findNodeHandle(this._componentRef.current),
-        this._getViewManagerConfig('RCTPSPDFKitView').Commands.getAnnotationFlags,
+        this._getViewManagerConfig('RCTPSPDFKitView').Commands
+          .getAnnotationFlags,
         [requestId, uuid],
       );
 
@@ -739,7 +764,7 @@ class PSPDFKitView extends React.Component {
    * Imports the supplied XFDF file into the current document.
    *
    * @method importXFDF
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.importXFDF()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().importXFDF()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.importXFDF|importXFDF()}.
    * @memberof PSPDFKitView
    * @param { string } filePath The path to the XFDF file to import.
@@ -777,7 +802,7 @@ class PSPDFKitView extends React.Component {
    * Exports the annotations from the current document to a XFDF file.
    *
    * @method exportXFDF
-   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument()?.exportXFDF()``` instead.
+   * @deprecated Since PSPDFKit for React Native 2.12. Use ```this.pdfRef.current?.getDocument().exportXFDF()``` instead.
    * See {@link https://pspdfkit.com/api/react-native/PDFDocument.html#.exportXFDF|exportXFDF()}.
    * @memberof PSPDFKitView
    * @param { string } filePath The path where the XFDF file should be exported to.
@@ -1022,21 +1047,20 @@ class PSPDFKitView extends React.Component {
    *	this.pdfRef.current.setToolbar(toolbar);
    *
    */
-    setToolbar = function (toolbar) {
-      if (Platform.OS === 'ios') {
-        NativeModules.PSPDFKitViewManager.setToolbar(
-          toolbar,
-          findNodeHandle(this._componentRef.current),
-        );
-      } else if (Platform.OS === 'android') {
-        UIManager.dispatchViewManagerCommand(
-          findNodeHandle(this._componentRef.current),
-          this._getViewManagerConfig('RCTPSPDFKitView').Commands
-          .setToolbar,
-          [toolbar],
-        );
-      }
-    };
+  setToolbar = function (toolbar) {
+    if (Platform.OS === 'ios') {
+      NativeModules.PSPDFKitViewManager.setToolbar(
+        toolbar,
+        findNodeHandle(this._componentRef.current),
+      );
+    } else if (Platform.OS === 'android') {
+      UIManager.dispatchViewManagerCommand(
+        findNodeHandle(this._componentRef.current),
+        this._getViewManagerConfig('RCTPSPDFKitView').Commands.setToolbar,
+        [toolbar],
+      );
+    }
+  };
 
   /**
    * Gets the toolbar for the specified view mode.
@@ -1050,14 +1074,14 @@ class PSPDFKitView extends React.Component {
    * const toolbar = await this.pdfRef.current.getToolbar('document');
    *
    */
-    getToolbar = function (viewMode) {
-      if (Platform.OS === 'ios') {
-        return NativeModules.PSPDFKitViewManager.getToolbar(
-          viewMode,
-          findNodeHandle(this._componentRef.current),
-        );
-      }
-    };
+  getToolbar = function (viewMode) {
+    if (Platform.OS === 'ios') {
+      return NativeModules.PSPDFKitViewManager.getToolbar(
+        viewMode,
+        findNodeHandle(this._componentRef.current),
+      );
+    }
+  };
 
   /**
    * Sets the measurement value configurations for the ```PSPDFKitView```.
@@ -1072,13 +1096,13 @@ class PSPDFKitView extends React.Component {
    *    unitTo: Measurements.ScaleUnitTo.INCH,
    *    valueTo: 2.54
    *  };
-   *  
+   *
    *  const measurementValueConfig: MeasurementValueConfiguration = {
    *    name: 'Custom Scale',
    *    scale: scale,
    *    precision: Measurements.Precision.FOUR_DP
    *  };
-   *  
+   *
    *  const configs = [measurementValueConfig];
    *  await this.pdfRef.current?.setMeasurementValueConfigurations(configs);
    */
@@ -1094,9 +1118,8 @@ class PSPDFKitView extends React.Component {
 
       UIManager.dispatchViewManagerCommand(
         findNodeHandle(this._componentRef.current),
-        this._getViewManagerConfig(
-          'RCTPSPDFKitView',
-        ).Commands.setMeasurementValueConfigurations,
+        this._getViewManagerConfig('RCTPSPDFKitView').Commands
+          .setMeasurementValueConfigurations,
         [requestId, configurations],
       );
 
@@ -1137,14 +1160,13 @@ class PSPDFKitView extends React.Component {
       );
 
       return promise;
-    }
-    else if (Platform.OS === 'ios') {
+    } else if (Platform.OS === 'ios') {
       return NativeModules.PSPDFKitViewManager.getMeasurementValueConfigurations(
         findNodeHandle(this._componentRef.current),
       );
     }
   };
-  
+
   /**
    * Customizes the visible toolbar menu items for Android.
    *
@@ -1189,14 +1211,12 @@ class PSPDFKitView extends React.Component {
 
       UIManager.dispatchViewManagerCommand(
         findNodeHandle(this._componentRef.current),
-        this._getViewManagerConfig('RCTPSPDFKitView').Commands
-          .getConfiguration,
+        this._getViewManagerConfig('RCTPSPDFKitView').Commands.getConfiguration,
         [requestId],
       );
 
       return promise;
-    }
-    else if (Platform.OS === 'ios') {
+    } else if (Platform.OS === 'ios') {
       return NativeModules.PSPDFKitViewManager.getConfiguration(
         findNodeHandle(this._componentRef.current),
       );
@@ -1290,7 +1310,7 @@ if (Platform.OS === 'ios' || Platform.OS === 'android') {
  * @property {string} [selectedFontName] Used to specify the current selected font in the font picker. Note on iOS: You need to set the desired font family names as ```UIFontDescriptor```. See {@link https://developer.apple.com/documentation/uikit/uifontdescriptor?language=objc} for more information. See {@link https://github.com/PSPDFKit/react-native/blob/master/samples/Catalog/examples/CustomFontPicker.tsx}
  * @property {boolean} [showDownloadableFonts] Used to show or hide the downloadable fonts section in the font picker. Defaults to ```true```, showing the downloadable fonts. See {@link https://developer.apple.com/documentation/uikit/uifontdescriptor?language=objc} for more information. See {@link https://github.com/PSPDFKit/react-native/blob/master/samples/Catalog/examples/CustomFontPicker.tsx}
  * @property {AnnotationPresetConfiguration} [annotationPresets] The annotation preset configuration. See {@link https://pspdfkit.com/api/react-native/Annotation.html#.AnnotationPresetConfiguration} for available options.
- * @property {boolean} [hideDefaultToolbar] Used to show or hide the annotation toolbar on Android.
+ * @property {boolean} [hideDefaultToolbar] Used to show or hide the main toolbar on Android.
  * @property {any} [style] Used to style the React Native component.
  *
  * @extends {Component<Props>}
@@ -1638,7 +1658,7 @@ PSPDFKitView.propTypes = {
    */
   annotationPresets: PropTypes.object,
   /**
-   * Used to show or hide the annotation toolbar on Android.
+   * Used to show or hide the main toolbar on Android.
    * Defaults to ```true```, showing the toolbar.
    * @type {boolean}
    * @memberof PSPDFKitView
@@ -1680,12 +1700,13 @@ export class PSPDFKit {
   versionString;
 
   /**
-   * Used to set your PSPDFKit license key for the iOS platform only. PSPDFKit is commercial software.
+   * Used to set your PSPDFKit license key for the active platform only, either iOS or Android.
+   * PSPDFKit is commercial software.
    * Each PSPDFKit license is bound to a specific app bundle ID.
    * Visit {@link https://customers.pspdfkit.com} to get your demo or commercial license key.
    * @method setLicenseKey
    * @memberof PSPDFKit
-   * @param { string | null } [key] Your PSPDFKit for React Native iOS license key.
+   * @param { string | null } [key] Your PSPDFKit for React Native iOS or PSPDFKit for React Native Android license key.
    * @returns { Promise<boolean> } A promise returning ```true``` if the license key was set, and ```false``` if not.
    * @example
    * PSPDFKit.setLicenseKey('YOUR_LICENSE_KEY');
@@ -1693,7 +1714,8 @@ export class PSPDFKit {
   setLicenseKey = function (key) {};
 
   /**
-   * Used to set the your PSPDFKit license keys for both platforms. PSPDFKit is commercial software.
+   * Used to set the your PSPDFKit license keys for both platforms.
+   * PSPDFKit is commercial software.
    * Each PSPDFKit license is bound to a specific app bundle ID.
    * Visit {@link https://customers.pspdfkit.com} to get your demo or commercial license key.
    * @method setLicenseKeys
@@ -1771,12 +1793,12 @@ export class PSPDFKit {
    *                      password);
    */
   processAnnotations = function (
-      annotationChange,
-      annotationTypes,
-      sourceDocumentPath,
-      processedDocumentPath,
-      password
-    ) {};
+    annotationChange,
+    annotationTypes,
+    sourceDocumentPath,
+    processedDocumentPath,
+    password,
+  ) {};
 
   /**
    * Used to present an Instant PDF document for collaboration.
@@ -2087,37 +2109,54 @@ export class Processor {
   getTemporaryDirectory = function () {};
 }
 
-import { PDFConfiguration } from "./lib/configuration/PDFConfiguration";
-export { PDFConfiguration } from "./lib/configuration/PDFConfiguration";
+import { PDFConfiguration } from './lib/configuration/PDFConfiguration';
+export { PDFConfiguration } from './lib/configuration/PDFConfiguration';
 
-import { RemoteDocumentConfiguration } from "./lib/configuration/PDFConfiguration";
-export { RemoteDocumentConfiguration } from "./lib/configuration/PDFConfiguration";
+import { RemoteDocumentConfiguration } from './lib/configuration/PDFConfiguration';
+export { RemoteDocumentConfiguration } from './lib/configuration/PDFConfiguration';
 
-import { Toolbar } from "./lib/toolbar/Toolbar";
-export { Toolbar } from "./lib/toolbar/Toolbar";
+import { Toolbar } from './lib/toolbar/Toolbar';
+export { Toolbar } from './lib/toolbar/Toolbar';
 
-import { Measurements } from "./lib/measurements/Measurements";
-export { Measurements } from "./lib/measurements/Measurements";
+import { Measurements } from './lib/measurements/Measurements';
+export { Measurements } from './lib/measurements/Measurements';
 
-import { MeasurementScale } from "./lib/measurements/Measurements";
-export { MeasurementScale } from "./lib/measurements/Measurements";
+import { MeasurementScale } from './lib/measurements/Measurements';
+export { MeasurementScale } from './lib/measurements/Measurements';
 
-import { MeasurementValueConfiguration } from "./lib/measurements/Measurements";
-export { MeasurementValueConfiguration } from "./lib/measurements/Measurements";
+import { MeasurementValueConfiguration } from './lib/measurements/Measurements';
+export { MeasurementValueConfiguration } from './lib/measurements/Measurements';
 
-import { Annotation } from "./lib/annotations/Annotation";
-export { Annotation } from "./lib/annotations/Annotation";
+import { Annotation } from './lib/annotations/Annotation';
+export { Annotation } from './lib/annotations/Annotation';
 
-import { AnnotationContextualMenu } from "./lib/annotations/Annotation";
-export { AnnotationContextualMenu } from "./lib/annotations/Annotation";
+import { AnnotationContextualMenu } from './lib/annotations/Annotation';
+export { AnnotationContextualMenu } from './lib/annotations/Annotation';
 
-import { AnnotationContextualMenuItem } from "./lib/annotations/Annotation";
-export { AnnotationContextualMenuItem } from "./lib/annotations/Annotation";
+import { AnnotationContextualMenuItem } from './lib/annotations/Annotation';
+export { AnnotationContextualMenuItem } from './lib/annotations/Annotation';
 
-import { AnnotationPresetConfiguration } from "./lib/annotations/Annotation";
-export { AnnotationPresetConfiguration } from "./lib/annotations/Annotation";
+import { AnnotationPresetConfiguration } from './lib/annotations/Annotation';
+export { AnnotationPresetConfiguration } from './lib/annotations/Annotation';
 
-import { AnnotationPresetInk,
+import {
+  AnnotationPresetEraser,
+  AnnotationPresetFile,
+  AnnotationPresetFreeText,
+  AnnotationPresetInk,
+  AnnotationPresetLine,
+  AnnotationPresetMarkup,
+  AnnotationPresetMeasurementArea,
+  AnnotationPresetMeasurementDistance,
+  AnnotationPresetMeasurementPerimeter,
+  AnnotationPresetNote,
+  AnnotationPresetRedact,
+  AnnotationPresetShape,
+  AnnotationPresetSound,
+  AnnotationPresetStamp,
+} from './lib/annotations/Annotation';
+export {
+  AnnotationPresetInk,
   AnnotationPresetFreeText,
   AnnotationPresetStamp,
   AnnotationPresetNote,
@@ -2130,24 +2169,14 @@ import { AnnotationPresetInk,
   AnnotationPresetRedact,
   AnnotationPresetMeasurementArea,
   AnnotationPresetMeasurementPerimeter,
-  AnnotationPresetMeasurementDistance } from "./lib/annotations/Annotation";
-export { AnnotationPresetInk,
-  AnnotationPresetFreeText,
-  AnnotationPresetStamp,
-  AnnotationPresetNote,
-  AnnotationPresetMarkup,
-  AnnotationPresetShape,
-  AnnotationPresetLine,
-  AnnotationPresetEraser,
-  AnnotationPresetFile,
-  AnnotationPresetSound,
-  AnnotationPresetRedact,
-  AnnotationPresetMeasurementArea,
-  AnnotationPresetMeasurementPerimeter,
-  AnnotationPresetMeasurementDistance } from "./lib/annotations/Annotation";
+  AnnotationPresetMeasurementDistance,
+} from './lib/annotations/Annotation';
 
-import { PDFDocument } from "./lib/document/PDFDocument";
-export { PDFDocument } from "./lib/document/PDFDocument";
+import { PDFDocument } from './lib/document/PDFDocument';
+export { PDFDocument } from './lib/document/PDFDocument';
+
+import { NotificationCenter } from './lib/notification-center/NotificationCenter';
+export { NotificationCenter } from './lib/notification-center/NotificationCenter';
 
 module.exports.PDFConfiguration = PDFConfiguration;
 module.exports.RemoteDocumentConfiguration = RemoteDocumentConfiguration;
@@ -2160,19 +2189,23 @@ module.exports.AnnotationContextualMenu = AnnotationContextualMenu;
 module.exports.AnnotationContextualMenuItem = AnnotationContextualMenuItem;
 module.exports.AnnotationPresetConfiguration = AnnotationPresetConfiguration;
 
-module.exports.AnnotationPresetInk = AnnotationPresetInk
-module.exports.AnnotationPresetFreeText = AnnotationPresetInk
-module.exports.AnnotationPresetStamp = AnnotationPresetStamp
-module.exports.AnnotationPresetNote = AnnotationPresetNote
-module.exports.AnnotationPresetMarkup = AnnotationPresetMarkup
-module.exports.AnnotationPresetShape = AnnotationPresetShape
-module.exports.AnnotationPresetLine = AnnotationPresetLine
-module.exports.AnnotationPresetEraser = AnnotationPresetEraser
-module.exports.AnnotationPresetFile = AnnotationPresetFile
-module.exports.AnnotationPresetSound = AnnotationPresetSound
-module.exports.AnnotationPresetRedact = AnnotationPresetRedact
-module.exports.AnnotationPresetMeasurementArea = AnnotationPresetMeasurementArea
-module.exports.AnnotationPresetMeasurementPerimeter = AnnotationPresetMeasurementPerimeter
-module.exports.AnnotationPresetMeasurementDistance = AnnotationPresetMeasurementDistance
+module.exports.AnnotationPresetInk = AnnotationPresetInk;
+module.exports.AnnotationPresetFreeText = AnnotationPresetFreeText;
+module.exports.AnnotationPresetStamp = AnnotationPresetStamp;
+module.exports.AnnotationPresetNote = AnnotationPresetNote;
+module.exports.AnnotationPresetMarkup = AnnotationPresetMarkup;
+module.exports.AnnotationPresetShape = AnnotationPresetShape;
+module.exports.AnnotationPresetLine = AnnotationPresetLine;
+module.exports.AnnotationPresetEraser = AnnotationPresetEraser;
+module.exports.AnnotationPresetFile = AnnotationPresetFile;
+module.exports.AnnotationPresetSound = AnnotationPresetSound;
+module.exports.AnnotationPresetRedact = AnnotationPresetRedact;
+module.exports.AnnotationPresetMeasurementArea =
+  AnnotationPresetMeasurementArea;
+module.exports.AnnotationPresetMeasurementPerimeter =
+  AnnotationPresetMeasurementPerimeter;
+module.exports.AnnotationPresetMeasurementDistance =
+  AnnotationPresetMeasurementDistance;
 
-module.exports.PDFDocument = PDFDocument;
+module.exports.NotificationCenter = NotificationCenter;
+module.exports.NotificationCenter = NotificationCenter;