+++ src/configuration/PDFConfiguration.ts
@@ -101,6 +101,9 @@ import { MeasurementValueConfiguration } from './../measurements/Measurements';
  * @property { PDFConfiguration.BooleanType } [syncAnnotations] Indicates whether document annotations should be synced with the Instant server.
  * @property { Measurements.MeasurementValueConfiguration[] } [measurementValueConfigurations] The array of ```MeasurementValueConfiguration``` objects that should be applied to the document.
  * @property { PDFConfiguration.RemoteDocumentConfiguration } [remoteDocumentConfiguration] The configuration when downloading a document from a remote URL.
+ * @property { PDFConfiguration.BooleanType } [androidShowDefaultToolbar] Used to show or hide the main toolbar.
+ * @property { PDFConfiguration.BooleanType } [showActionButtons] Shows floating back and forward buttons in the lower part of the screen.
+ * 
  */
 
 export class PDFConfiguration {
@@ -475,6 +478,16 @@ export class PDFConfiguration {
      * The configuration when downloading a document from a remote URL.
      */
     remoteDocumentConfiguration?: RemoteDocumentConfiguration;
+
+    /**
+     * Used to show or hide the main toolbar.
+     */
+    androidShowDefaultToolbar?: PDFConfiguration.BooleanType;
+
+    /**
+     * Shows floating back and forward buttons in the lower part of the screen.
+     */
+    showActionButtons?: PDFConfiguration.BooleanType;
 }
 
 export namespace PDFConfiguration {